Implement environment monad
EnvM is a ReaderT stack over IO which carries a cache of the .conf files
and options passed in from the command line. This should help reduce
unnecessary I/O and remove the need to explicitly pass Verbosity around.

The 'say' and 'vsay' functions have been moved to MonadSay, which allows
for EnvM and the simpler SayM to use both easily.

Signed-off-by: hololeap <hololeap@users.noreply.github.com>