Main: Add improvements to runUpdater
Change runUpdater to run inside a MonadState which holds the package
state and history. Add UpdateState type synonym.

Rework looping logic to be less error prone. Change UpdaterLoop so that
it is a function, which takes the next iteration of the loop as an
argument. Add runLoop which calls the loop using
`updateAndContinue loop` as an argument.

Make runUpdater more generic, taking various constraints instead of
using Env.

Move system info display, list mode logic, and gathering of initial
state to runAction.

Signed-off-by: hololeap <hololeap@protonmail.com>