Main: Fix loopUntilNoChange logic
The logic is applied _after_ a run is finished, so we need to store the
initial package state in RunHistory. This allows us to compare the
state after the first run with the initial state to see if anything
changed.

Add IOEnv newtype wrapper to Main, which encapsulates the normal StateT
over IO monad transformer stack. This helps the compiler make sense of
the type loop that forms from (UpdaterState (StateT (UpdaterState ...)))

Signed-off-by: hololeap <hololeap@protonmail.com>